Colocate fleet table wire parser with its Zod schema module.
Remove the pass-through parseFleetTableWire wrapper file; parsing and validation errors live alongside the schema definitions.
